Fix | CustomDB.js, EnumTableName.js : 테이블 리터널 분리
테이블 리터널 내용들을 EnumTableName으로 분리하였음
